pz_tracker/requirements.txt
playwright==1.49.1
pdfplumber==0.11.4
PyMuPDF==1.25.1
groq>=1.1.0
openpyxl==3.1.5
python-dotenv==1.0.1
httpx==0.28.1
schedule==1.2.2
